pyproject.toml
[tool.poetry]
name = "coursework-7"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
celery = "^5.3.6"
python-dotenv = "^1.0.1"
drf-yasg = "^1.21.7"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.2"
django-celery-beat = "^2.6.0"
psycopg2 = "^2.9.9"
pillow = "^10.3.0"
django-cors-headers = "^4.3.1"
redis = "^5.0.3"
flake8 = "^7.0.0"
requests = "^2.31.0"
python-crontab = "^3.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
